Change data type so it can hold val of 256

+++ rejistry++/src/NKRecord.h
@@ -167,7 +167,7 @@ namespace Rejistry {
         // Microsoft docs state that max key name is 255 characters. This is true when testing
         // with regedit.exe and reg.exe. However, when using RegCreateKeyEx() API you can create
         // keys with 256 characters. See CT-2984 for test data. 
-        static const uint8_t MAX_NAME_LENGTH = 256; //# of characters
+        static const uint16_t MAX_NAME_LENGTH = 256; //# of characters
 
         NKRecord() {};
         NKRecord& operator=(const NKRecord &);
